Split flow-and-prod-dev-entry plugin up
